+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Performance improvements for $addToSet update operator up to ~70x for large existing and update arrays. *[Perf]*
 * Removed feature flags `documentdb.enableCompact`, `documentdb.enableBucketAutoStage` and `documentdb.enableIndexHintSupport`
 * Fix use-after-free segmentation fault in `$let` *[Bugfix]* 
+* Short-circuit in `$switch` at parse time *[Perf]*
 * Enable ordered indexes by default. Can be turned off by specifying "storageEngine": {"enableOrderedIndex": false} for a single index or by turning off the `documentdb.defaultUseCompositeOpClass` GUC.
 
 ### documentdb v0.108-0 (Unreleased) ###